feat: Refactor InventorySection and remove SearchSection; add AI features and dynamic research components
- Updated InventorySection to use cardBackground instead of paperBackground for better theming.
- Increased autoplay delay in InventorySection carousel from 3000ms to 4000ms.
- Removed SearchSection component to streamline the UI.
- Introduced new AI features component for enhanced user interaction with AI-powered medicine search and prescription scanning.
- Added dynamic research component to improve search functionality and user experience.
- Implemented animated trust indicators to build user confidence in the platform.
- Enhanced hero section with new animations and visual elements for a more engaging experience.
